// src/main/generated/typescript-helpers.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type Component = {
  __typename?: 'Component';
  key: Scalars['String'];
  location: Scalars['String'];
  parameters?: Maybe<Scalars['String']>;
  sortOrder: ComponentOrder;
  type: Scalars['String'];
  viewKey: View;
};

export type ComponentOrder = {
  __typename?: 'ComponentOrder';
  componentKey: Scalars['String'];
  sortOrder: Scalars['Int'];
};
// ...
export type View = {
  __typename?: 'View';
  createdAt?: Maybe<Scalars['DateTime']>;
  deleted: Scalars['Boolean'];
  deletedAt?: Maybe<Scalars['DateTime']>;
  icon: Scalars['String'];
  key: Scalars['String'];
  name: Scalars['String'];
  sortOrder?: Maybe<ViewOrder>;
  type: Scalars['String'];
};

export type ViewOrder = {
  __typename?: 'ViewOrder';
  sortOrder: Scalars['Int'];
  viewKey: Scalars['String'];
};